+++ src/Stratis.Bitcoin/Base/AsyncProvider/AsyncLoop.cs
