+++ src/Stratis.Bitcoin/Base/AsyncProvider/AsyncLoopFactory.cs
