+++ src/Stratis.Bitcoin/Base/AsyncProvider/IAsyncQueue.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Stratis.Bitcoin.Base.BackgroundWork
+namespace Stratis.Bitcoin.Base.AsyncProvider
 {
     /// <summary>
     /// Defines a queue that can be dequeued asynchronously by multiple threads.
