+++ src/Stratis.Bitcoin/Base/BaseFeature.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
-using Stratis.Bitcoin.Base.BackgroundWork;
+using Stratis.Bitcoin.Base.AsyncProvider;
 using Stratis.Bitcoin.Base.Deployments;
 using Stratis.Bitcoin.BlockPulling;
 using Stratis.Bitcoin.Builder;
@@ -384,7 +384,7 @@ namespace Stratis.Bitcoin.Base
                     services.AddSingleton<IInitialBlockDownloadState, InitialBlockDownloadState>();
                     services.AddSingleton<IKeyValueRepository, KeyValueRepository>();
                     services.AddSingleton<ITipsManager, TipsManager>();
-                    services.AddSingleton<IBackgroundWorkProvider, BackgroundWorkProvider>();
+                    services.AddSingleton<IBackgroundWorkProvider, AsyncProvider>();
 
                     // Consensus
                     services.AddSingleton<ConsensusSettings>();
